Refactor/simplify TideChart
