+++ packages/react/src/components/TideGraph/constants.ts
@@ -1,0 +1,6 @@
+import type { Margin } from "../../utils/scales.js";
+
+export const PX_PER_DAY_DEFAULT = 200;
+export const HEIGHT = 300;
+export const MARGIN: Margin = { top: 65, right: 0, bottom: 40, left: 60 };
+export const MS_PER_DAY = 24 * 60 * 60 * 1000;
